timeout almost done, needs to pass some more tests

+++ signals.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <signal.h>
 #include "signals.h"
+#include <sys/wait.h>
 #include "Commands.h"
 
 using namespace std;
@@ -22,9 +23,19 @@ void ctrlCHandler(int sig_num) {
 
 void alarmHandler(int sig_num) {
   // TODO: Add your implementation
+  std::cout<<"smash: got an alarm"<<std::endl;
   JobsList::TimedJob job = SmallShell::getInstance().getTimedJobs().top();
-  SmallShell::getInstance().getTimedJobs().pop();
-  kill(job.get_pid(),SIGKILL);
-  std::cout<<"smash: "<<job.getCommand()<<" timed out!";
+  if (!waitpid(job.get_pid(), nullptr, WNOHANG))
+  {
+    SmallShell::getInstance().getTimedJobs().pop();
+    kill(job.get_pid(), SIGKILL);
+    std::cout << "smash: " << job.getCommand() << " timed out!"<<std::endl;
+    if (!SmallShell::getInstance().getTimedJobs().empty()){
+      int top_original_alarm_time = SmallShell::getInstance().getTimedJobs().top().getAlarmTime();
+      int top_already_passed_time = time(nullptr)-SmallShell::getInstance().getTimedJobs().top().getTimeMade();
+      alarm(top_original_alarm_time-top_already_passed_time);
+    }
+    
+  }
 }
